Add each helper function in a new file
